Change namespace, now with option -k I am able to see the response using curl
