+++ src/main/java/https/somenamespace/ws/GetObjectRequest.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.01.28 at 01:24:18 AM EET 
+// Generated on: 2018.01.28 at 09:53:24 AM EET 
 //
 
 
-package io.spring.guides.gs_producing_web_service;
+package https.somenamespace.ws;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -25,7 +25,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="object" type="{http://spring.io/guides/gs-producing-web-service}someRandomObject"/>
+ *         &lt;element name="userName" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -36,36 +36,36 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "object"
+    "userName"
 })
-@XmlRootElement(name = "getObjectResponse")
-public class GetObjectResponse {
+@XmlRootElement(name = "getObjectRequest")
+public class GetObjectRequest {
 
     @XmlElement(required = true)
-    protected SomeRandomObject object;
+    protected String userName;
 
     /**
-     * Gets the value of the object property.
+     * Gets the value of the userName property.
      * 
      * @return
      *     possible object is
-     *     {@link SomeRandomObject }
+     *     {@link String }
      *     
      */
-    public SomeRandomObject getObject() {
-        return object;
+    public String getUserName() {
+        return userName;
     }
 
     /**
-     * Sets the value of the object property.
+     * Sets the value of the userName property.
      * 
      * @param value
      *     allowed object is
-     *     {@link SomeRandomObject }
+     *     {@link String }
      *     
      */
-    public void setObject(SomeRandomObject value) {
-        this.object = value;
+    public void setUserName(String value) {
+        this.userName = value;
     }
 
 }
