+++ src/main/java/https/somenamespace/ws/ObjectFactory.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.01.28 at 01:24:18 AM EET 
+// Generated on: 2018.01.28 at 09:53:24 AM EET 
 //
 
 
-package io.spring.guides.gs_producing_web_service;
+package https.somenamespace.ws;
 
 import javax.xml.bind.annotation.XmlRegistry;
 
@@ -14,7 +14,7 @@ import javax.xml.bind.annotation.XmlRegistry;
 /**
  * This object contains factory methods for each 
  * Java content interface and Java element interface 
- * generated in the io.spring.guides.gs_producing_web_service package. 
+ * generated in the https.somenamespace.ws package. 
  * <p>An ObjectFactory allows you to programatically 
  * construct new instances of the Java representation 
  * for XML content. The Java representation of XML 
@@ -30,7 +30,7 @@ public class ObjectFactory {
 
 
     /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: io.spring.guides.gs_producing_web_service
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: https.somenamespace.ws
      * 
      */
     public ObjectFactory() {
