+++ src/main/java/https/somenamespace/ws/SomeRandomObject.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.01.28 at 01:24:18 AM EET 
+// Generated on: 2018.01.28 at 09:53:24 AM EET 
 //
 
 
-package io.spring.guides.gs_producing_web_service;
+package https.somenamespace.ws;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
