+++ src/main/java/https/somenamespace/ws/package-info.java
@@ -2,8 +2,8 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2018.01.28 at 01:24:18 AM EET 
+// Generated on: 2018.01.28 at 09:53:24 AM EET 
 //
 
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://spring.io/guides/gs-producing-web-service", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package io.spring.guides.gs_producing_web_service;
+@javax.xml.bind.annotation.XmlSchema(namespace = "https://somenamespace/ws", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+package https.somenamespace.ws;
