+++ src/main/resources/example.xsd
@@ -1,5 +1,5 @@
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns:tns="http://spring.io/guides/gs-producing-web-service"
-           targetNamespace="http://spring.io/guides/gs-producing-web-service" elementFormDefault="qualified">
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns:tns="https://somenamespace/ws"
+           targetNamespace="https://somenamespace/ws" elementFormDefault="qualified">
 
     <xs:element name="getObjectRequest">
         <xs:complexType>
