+++ src/main/resources/request.xml
@@ -1,5 +1,5 @@
 <soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/"
-                  xmlns:gs="http://spring.io/guides/gs-producing-web-service">
+                  xmlns:gs="https://somenamespace/ws">
     <soapenv:Header/>
     <soapenv:Body>
         <gs:getObjectRequest>
